requirements.txt
jinja2
tqdm
absl-py
openai
pyyaml
google-cloud-aiplatform
anthropic
marko
pandas